Split E2B tools into dedicated files per class
- e2b_sandbox.py: E2BSandbox
- bash_tool.py: RunBashCommandTool
- read_file_tool.py: ReadFileTool
- write_file_tool.py: WriteFileTool
- list_directory_tool.py: ListDirectoryTool
- sandbox_toolset.py: create_e2b_tools (convenience function only)
