+++ haystack_experimental/tools/e2b/__init__.py
@@ -8,23 +8,21 @@ from typing import TYPE_CHECKING
 from lazy_imports import LazyImporter
 
 _import_structure = {
-    "sandbox_toolset": [
-        "E2BSandbox",
-        "RunBashCommandTool",
-        "ReadFileTool",
-        "WriteFileTool",
-        "ListDirectoryTool",
-        "create_e2b_tools",
-    ]
+    "e2b_sandbox": ["E2BSandbox"],
+    "bash_tool": ["RunBashCommandTool"],
+    "read_file_tool": ["ReadFileTool"],
+    "write_file_tool": ["WriteFileTool"],
+    "list_directory_tool": ["ListDirectoryTool"],
+    "sandbox_toolset": ["create_e2b_tools"],
 }
 
 if TYPE_CHECKING:
-    from .sandbox_toolset import E2BSandbox as E2BSandbox
-    from .sandbox_toolset import ListDirectoryTool as ListDirectoryTool
-    from .sandbox_toolset import ReadFileTool as ReadFileTool
-    from .sandbox_toolset import RunBashCommandTool as RunBashCommandTool
-    from .sandbox_toolset import WriteFileTool as WriteFileTool
+    from .bash_tool import RunBashCommandTool as RunBashCommandTool
+    from .e2b_sandbox import E2BSandbox as E2BSandbox
+    from .list_directory_tool import ListDirectoryTool as ListDirectoryTool
+    from .read_file_tool import ReadFileTool as ReadFileTool
     from .sandbox_toolset import create_e2b_tools as create_e2b_tools
+    from .write_file_tool import WriteFileTool as WriteFileTool
 
 else:
     sys.modules[__name__] = LazyImporter(name=__name__, module_file=__file__, import_structure=_import_structure)
