+++ haystack_experimental/tools/e2b/bash_tool.py
@@ -1,0 +1,68 @@
+# SPDX-FileCopyrightText: 2022-present deepset GmbH <info@deepset.ai>
+#
+# SPDX-License-Identifier: Apache-2.0
+
+from haystack.tools import Tool
+
+from haystack_experimental.tools.e2b.e2b_sandbox import E2BSandbox
+
+
+class RunBashCommandTool(Tool):
+    """
+    A :class:`~haystack.tools.Tool` that executes bash commands inside an E2B sandbox.
+
+    Pass the same :class:`E2BSandbox` instance to multiple tool classes so they
+    all operate in the same live sandbox environment.
+
+    ### Usage example
+
+    ```python
+    from haystack_experimental.tools.e2b import E2BSandbox, RunBashCommandTool, ReadFileTool
+
+    sandbox = E2BSandbox()
+    agent = Agent(
+        chat_generator=...,
+        tools=[
+            RunBashCommandTool(sandbox=sandbox),
+            ReadFileTool(sandbox=sandbox),
+        ],
+    )
+    ```
+    """
+
+    def __init__(self, sandbox: E2BSandbox) -> None:
+        """
+        :param sandbox: The :class:`E2BSandbox` instance that will execute commands.
+        """
+
+        def run_bash_command(command: str, timeout: int = 60) -> str:
+            sb = sandbox._require_sandbox()
+            try:
+                result = sb.commands.run(command, timeout=timeout)
+                return f"exit_code: {result.exit_code}\nstdout:\n{result.stdout}\nstderr:\n{result.stderr}"
+            except Exception as e:
+                raise RuntimeError(f"Failed to run bash command: {e}") from e
+
+        super().__init__(
+            name="run_bash_command",
+            description=(
+                "Execute a bash command inside the E2B sandbox and return the combined stdout, "
+                "stderr, and exit code. Use this to run shell scripts, install packages, compile "
+                "code, or perform any system-level operation."
+            ),
+            parameters={
+                "type": "object",
+                "properties": {
+                    "command": {"type": "string", "description": "The bash command to execute."},
+                    "timeout": {
+                        "type": "integer",
+                        "description": (
+                            "Maximum number of seconds to wait for the command to finish. Defaults to 60 seconds."
+                        ),
+                    },
+                },
+                "required": ["command"],
+            },
+            function=run_bash_command,
+        )
+        self._e2b_sandbox = sandbox
